Adding patch for Telia IMS registration/call dependency
from svn.digium.com/svn/asterisk/team/oej/register-call-dependency-1.8

In order to place a call, we need to call the very same server we
registred to. In this patch we solve that with a new config setting
(to be documented) and by using the outbound proxy of the peer.
When we register, we need to refer to the peer in the register=
line and NOT configure an outbound proxy for the peer (it will
simply be overwritten).

This patch does not change any behaviour in regards to DNS
srv records